Prefer u64 or u32 over usize.
Using `usize` in types that get serialized can cause compatibility
issues between different architectures.

+++ linera-core/src/data_types.rs
@@ -31,7 +31,7 @@ pub struct BlockHeightRange {
     /// Starting point
     pub start: BlockHeight,
     /// Optional limit on the number of elements.
-    pub limit: Option<usize>,
+    pub limit: Option<u64>,
 }
 
 /// Message to obtain information on a chain.
